feat: refactor ComicReader to utilize utility functions for downloading and extracting CBZ archives
